dry up HasInputOptions which requires __ to be rebound
